+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.2.3",
   "description": "Tinkoff Invest OpenAPI Client",
   "main": "index.js",
+  "types": "./build/OpenAPI.d.ts",
   "directories": {
     "example": "example"
   },
@@ -25,6 +26,7 @@
   "license": "ISC",
   "devDependencies": {
     "@tinkoff/swagger-types-generator": "0.0.6",
+    "@types/ws": "^7.2.3",
     "fsevents": "^2.0.7",
     "prettier": "1.18.2",
     "rollup": "^1.21.3",
