+++ rollup.config.js
@@ -1,15 +1,6 @@
 import typescript from 'rollup-plugin-typescript2';
-import dts from 'rollup-plugin-dts';
 
 export default [
-  {
-    input: './src/domain.d.ts',
-    output: {
-      file: 'build/domain.d.ts',
-      format: 'es',
-    },
-    plugins: [dts()],
-  },
   {
     input: './src/OpenAPI.ts',
     output: {
